requirements.txt
Django==4.2.11
Pillow==10.2.0
pandas==2.1.4
openpyxl==3.1.2
